Rebuild the service.
Semikols (the previous version of this project) was supposed to be an
index of all my open source web projects, but I realised that I am not
planning on any other project in the nearest future, therefore there is
no pint in keeping such an index. Instead I ddicated this project to a
single service - Color Splash.